Realtime signal processing for powertrain dyno testing
A section of the implementation file for the real-time C++ program I developed during my time at Jaguar TCS Racing. It includes the methods available from the Kingstar api to setup and manage powertrain dyno hardware connected to the PC using the EtherCAT network protocol. It also includes the methods for scheduling the execution of other sections of the program not shown at high frequency (250 microseconds).